+++ CHANGELOG.md
@@ -6,6 +6,8 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
 ## [x.x.x]
 ### Added
+- Introduced a `reset dismiss variant` verb
+- Button to reset all dismissed variants for a case
 ### Fixed
 ### Changed
 
@@ -46,7 +48,6 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Submit to Beacon from case page sidebar
 - Hide dismissed variants in variants pages and variants export function
 - Systemd service files and instruction to deploy Scout using podman
-- Button to reset all dismissed variants for a case
 ### Fixed
 - Bugfix: unused `chromgraph_prefix |tojson` removed
 - Freeze coloredlogs temporarily